backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
pymysql>=1.1.0
python-dotenv>=1.0.1
httpx>=0.27.0
dnspython>=2.6.1
python-whois>=0.9.4
beautifulsoup4>=4.12.3
lxml>=5.1.0
requests>=2.31.0
python-multipart>=0.0.9
pydantic>=2.6.4
cryptography>=42.0.5
reportlab>=4.1.0
phonenumbers>=8.13.33
tldextract>=5.1.2
validators>=0.28.1
aiohttp>=3.9.4
